fix(gallery): show artworks via SSR merge + robust artwork queries
Artworks weren't appearing on the preview: the API merge could 500 (the
artwork sort over ~26k rows can exceed Mongo's 32MB in-memory limit), and the
client then silently fell back to the SSR illustration-only payload.

- Extract the merge to a pure, testable lib (src/lib/gallery-merge.ts); verified
  standalone it returns 18 illustrations + 6 artworks per page across offsets.
- Add allowDiskUse to the artwork sort so it can't blow the in-memory limit.
- SSR (fetchInitialGalleryData) now returns the MERGED feed for the plain
  gallery, so artworks show on first paint regardless of the client fetch.
- Artwork tiles render unoptimized (some artwork image hosts aren't in
  next/image remotePatterns, e.g. Leiden IIIF — would otherwise break).
